Multiple Fractal Regression

Calculates the scale-dependent (DFA-based) multiple linear regression: the coefficients, their variance and confidence intervals at each scale, together with scale-wise collinearity diagnostics.

Usage

fracreg(
  data,
  dpo,
  int,
  np = 91,
  overlap = TRUE,
  vcov = c("inverse", "marginal", "HC"),
  abs = FALSE
)

Arguments

data

a matrix or data frame of time series; the first column is the response and the remaining columns are the predictors.

dpo

detrending polynomial order.

int

logical. If TRUE the integration process is applied.

np

number of point scales.

overlap

logical. If TRUE overlapping windows are applied.

vcov

coefficient-variance estimator: "inverse" (default; full \(F_{XX}(s)^{-1}\), Tilfani et al. 2022), "marginal" (legacy \(1/F^2_{X_j}(s)\), Shen 2015), or "HC" (experimental heteroscedasticity-consistent sandwich built from the per-box detrended moment scores; overlap = FALSE is recommended and the finite-sample normalisation is still being calibrated).

abs

logical. If TRUE the absolute detrended covariance is used in the cross-correlation step (more robust to outliers).

Value

A list with, among others: scale s, detrended fluctuation F, DCCA, DPCCA, beta estimates BDFA, standardized betas BSDFA, residual variance UDFA, coefficient variance VDFA, multiple correlation DMC2, R2DFA, confidence limits UCIB/LCIB, p.value, critical value TC, and the scale-wise diagnostics VIF, condition number kappa and adjusted R2adj.

Details

The variance of the scale-dependent coefficients follows Tilfani et al. (2022, Eq. 25), \(\mathrm{var}(\hat\beta_j(s)) = F^2_\varepsilon(s) [F_{XX}(s)^{-1}]_{jj}\), i.e. it uses the full inverse of the detrended covariance matrix of the predictors, consistent with how the coefficients themselves are estimated. Under collinearity this differs from the legacy "marginal" form \(F^2_\varepsilon(s) / F^2_{X_j}(s)\) (which under-covers); the two coincide for orthogonal predictors. Set vcov = "marginal" to reproduce the legacy behaviour.

References

Barreto, I. D. C., Dore, L. H., Stosic, T. and Stosic, B. D. (2021). Extending DFA-based multiple linear regression inference: application to acoustic impedance models. Physica A, 582, 126259.

Tilfani, O., Kristoufek, L., Ferreira, P. and El Boukfaoui, M. Y. (2022). Heterogeneity in economic relationships: scale dependence through the multivariate fractal regression. Physica A, 588, 126530.

Shen, C. (2015). A new detrended semipartial cross-correlation analysis. Physics Letters A, 379(44), 2962-2969.

Kristoufek, L. (2015). Detrended fluctuation analysis as a regression framework: estimating dependence at different scales. Physical Review E, 91(2), 022802.

See also

Examples

set.seed(1)
d <- data.frame(y = cumsum(rnorm(300)), x1 = cumsum(rnorm(300)),
                x2 = cumsum(rnorm(300)))
fracreg(d, dpo = 1, int = TRUE, np = 20)
#> $s
#>  [1] 10 12 13 14 15 16 17 18 19 20 21 23 25 27 29 31 34 37 40 43
#> 
#> $F
#> , , 1
#> 
#>             [,1]       [,2]       [,3]
#> [1,]  1.93666212 0.09310289 -0.1228685
#> [2,]  0.09310289 2.42433776  0.3329354
#> [3,] -0.12286848 0.33293545  2.0107056
#> 
#> , , 2
#> 
#>            [,1]      [,2]       [,3]
#> [1,]  3.2932280 0.2239064 -0.2022399
#> [2,]  0.2239064 4.0324914  0.6878864
#> [3,] -0.2022399 0.6878864  3.2225709
#> 
#> , , 3
#> 
#>            [,1]      [,2]       [,3]
#> [1,]  4.1655660 0.3205775 -0.2631521
#> [2,]  0.3205775 5.0302202  0.9361710
#> [3,] -0.2631521 0.9361710  3.9775530
#> 
#> , , 4
#> 
#>            [,1]     [,2]       [,3]
#> [1,]  5.1866569 0.446654 -0.3442161
#> [2,]  0.4466540 6.163230  1.2398381
#> [3,] -0.3442161 1.239838  4.8545169
#> 
#> , , 5
#> 
#>            [,1]     [,2]       [,3]
#> [1,]  6.3743057 0.613541 -0.4482361
#> [2,]  0.6135410 7.438397  1.6065683
#> [3,] -0.4482361 1.606568  5.8744532
#> 
#> , , 6
#> 
#>            [,1]      [,2]       [,3]
#> [1,]  7.7402895 0.8319754 -0.5829193
#> [2,]  0.8319754 8.8630723  2.0438346
#> [3,] -0.5829193 2.0438346  7.0571651
#> 
#> , , 7
#> 
#>            [,1]      [,2]       [,3]
#> [1,]  9.2992535  1.112171 -0.7534072
#> [2,]  1.1121713 10.442110  2.5588207
#> [3,] -0.7534072  2.558821  8.4223151
#> 
#> , , 8
#> 
#>            [,1]      [,2]       [,3]
#> [1,] 11.0695604  1.463887 -0.9662741
#> [2,]  1.4638867 12.182940  3.1593359
#> [3,] -0.9662741  3.159336  9.9887859
#> 
#> , , 9
#> 
#>           [,1]      [,2]      [,3]
#> [1,] 13.070450  1.896721 -1.228722
#> [2,]  1.896721 14.094069  3.852461
#> [3,] -1.228722  3.852461 11.775118
#> 
#> , , 10
#> 
#>           [,1]      [,2]      [,3]
#> [1,] 15.321186  2.420875 -1.546786
#> [2,]  2.420875 16.186392  4.641688
#> [3,] -1.546786  4.641688 13.798479
#> 
#> , , 11
#> 
#>           [,1]      [,2]      [,3]
#> [1,] 17.838047  3.046478 -1.927889
#> [2,]  3.046478 18.468149  5.527422
#> [3,] -1.927889  5.527422 16.074707
#> 
#> , , 12
#> 
#>           [,1]      [,2]      [,3]
#> [1,] 23.726009  4.635707 -2.900819
#> [2,]  4.635707 23.625448  7.596123
#> [3,] -2.900819  7.596123 21.442725
#> 
#> , , 13
#> 
#>           [,1]      [,2]      [,3]
#> [1,] 30.833855  6.717495 -4.197098
#> [2,]  6.717495 29.632299 10.110653
#> [3,] -4.197098 10.110653 28.030048
#> 
#> , , 14
#> 
#>           [,1]     [,2]      [,3]
#> [1,] 39.290423  9.37445 -5.897794
#> [2,]  9.374450 36.56380 13.103754
#> [3,] -5.897794 13.10375 36.014768
#> 
#> , , 15
#> 
#>           [,1]     [,2]      [,3]
#> [1,] 49.233582 12.69212 -8.073894
#> [2,] 12.692117 44.44935 16.567067
#> [3,] -8.073894 16.56707 45.560020
#> 
#> , , 16
#> 
#>           [,1]     [,2]      [,3]
#> [1,]  60.81839 16.74967 -10.82605
#> [2,]  16.74967 53.30060  20.47970
#> [3,] -10.82605 20.47970  56.84408
#> 
#> , , 17
#> 
#>           [,1]     [,2]      [,3]
#> [1,]  81.52096 24.35036 -16.27012
#> [2,]  24.35036 68.38415  27.13020
#> [3,] -16.27012 27.13020  77.42360
#> 
#> , , 18
#> 
#>           [,1]     [,2]      [,3]
#> [1,] 106.48922 33.98771 -23.54945
#> [2,]  33.98771 85.69673  34.62027
#> [3,] -23.54945 34.62027 102.87978
#> 
#> , , 19
#> 
#>           [,1]      [,2]      [,3]
#> [1,] 136.04885  45.95149 -32.90508
#> [2,]  45.95149 105.31634  42.88488
#> [3,] -32.90508  42.88488 133.79985
#> 
#> , , 20
#> 
#>           [,1]      [,2]      [,3]
#> [1,] 170.43617  60.42402 -44.68286
#> [2,]  60.42402 127.30625  51.79230
#> [3,] -44.68286  51.79230 170.67600
#> 
#> 
#> $DCCA
#> , , 1
#> 
#>             [,1]       [,2]        [,3]
#> [1,]  1.00000000 0.04296746 -0.06226433
#> [2,]  0.04296746 1.00000000  0.15079579
#> [3,] -0.06226433 0.15079579  1.00000000
#> 
#> , , 2
#> 
#>             [,1]       [,2]        [,3]
#> [1,]  1.00000000 0.06144251 -0.06208042
#> [2,]  0.06144251 1.00000000  0.19082214
#> [3,] -0.06208042 0.19082214  1.00000000
#> 
#> , , 3
#> 
#>             [,1]       [,2]        [,3]
#> [1,]  1.00000000 0.07003297 -0.06464902
#> [2,]  0.07003297 1.00000000  0.20929251
#> [3,] -0.06464902 0.20929251  1.00000000
#> 
#> , , 4
#> 
#>             [,1]      [,2]        [,3]
#> [1,]  1.00000000 0.0789993 -0.06859848
#> [2,]  0.07899930 1.0000000  0.22666674
#> [3,] -0.06859848 0.2266667  1.00000000
#> 
#> , , 5
#> 
#>             [,1]      [,2]        [,3]
#> [1,]  1.00000000 0.0891020 -0.07324984
#> [2,]  0.08910200 1.0000000  0.24303891
#> [3,] -0.07324984 0.2430389  1.00000000
#> 
#> , , 6
#> 
#>             [,1]      [,2]        [,3]
#> [1,]  1.00000000 0.1004476 -0.07887052
#> [2,]  0.10044764 1.0000000  0.25842734
#> [3,] -0.07887052 0.2584273  1.00000000
#> 
#> , , 7
#> 
#>             [,1]      [,2]        [,3]
#> [1,]  1.00000000 0.1128635 -0.08513143
#> [2,]  0.11286350 1.0000000  0.27285378
#> [3,] -0.08513143 0.2728538  1.00000000
#> 
#> , , 8
#> 
#>             [,1]      [,2]        [,3]
#> [1,]  1.00000000 0.1260568 -0.09189223
#> [2,]  0.12605678 1.0000000  0.28639366
#> [3,] -0.09189223 0.2863937  1.00000000
#> 
#> , , 9
#> 
#>             [,1]      [,2]        [,3]
#> [1,]  1.00000000 0.1397462 -0.09904345
#> [2,]  0.13974618 1.0000000  0.29904558
#> [3,] -0.09904345 0.2990456  1.00000000
#> 
#> , , 10
#> 
#>            [,1]      [,2]       [,3]
#> [1,]  1.0000000 0.1537273 -0.1063821
#> [2,]  0.1537273 1.0000000  0.3105885
#> [3,] -0.1063821 0.3105885  1.0000000
#> 
#> , , 11
#> 
#>            [,1]      [,2]       [,3]
#> [1,]  1.0000000 0.1678467 -0.1138510
#> [2,]  0.1678467 1.0000000  0.3208037
#> [3,] -0.1138510 0.3208037  1.0000000
#> 
#> , , 12
#> 
#>            [,1]      [,2]       [,3]
#> [1,]  1.0000000 0.1958004 -0.1286081
#> [2,]  0.1958004 1.0000000  0.3374908
#> [3,] -0.1286081 0.3374908  1.0000000
#> 
#> , , 13
#> 
#>            [,1]      [,2]       [,3]
#> [1,]  1.0000000 0.2222341 -0.1427655
#> [2,]  0.2222341 1.0000000  0.3508202
#> [3,] -0.1427655 0.3508202  1.0000000
#> 
#> , , 14
#> 
#>            [,1]      [,2]       [,3]
#> [1,]  1.0000000 0.2473300 -0.1567855
#> [2,]  0.2473300 1.0000000  0.3611019
#> [3,] -0.1567855 0.3611019  1.0000000
#> 
#> , , 15
#> 
#>            [,1]      [,2]       [,3]
#> [1,]  1.0000000 0.2713130 -0.1704749
#> [2,]  0.2713130 1.0000000  0.3681467
#> [3,] -0.1704749 0.3681467  1.0000000
#> 
#> , , 16
#> 
#>            [,1]      [,2]       [,3]
#> [1,]  1.0000000 0.2941864 -0.1841237
#> [2,]  0.2941864 1.0000000  0.3720617
#> [3,] -0.1841237 0.3720617  1.0000000
#> 
#> , , 17
#> 
#>            [,1]      [,2]       [,3]
#> [1,]  1.0000000 0.3261318 -0.2047950
#> [2,]  0.3261318 1.0000000  0.3728538
#> [3,] -0.2047950 0.3728538  1.0000000
#> 
#> , , 18
#> 
#>            [,1]      [,2]       [,3]
#> [1,]  1.0000000 0.3557844 -0.2249899
#> [2,]  0.3557844 1.0000000  0.3687087
#> [3,] -0.2249899 0.3687087  1.0000000
#> 
#> , , 19
#> 
#>            [,1]      [,2]       [,3]
#> [1,]  1.0000000 0.3838878 -0.2438865
#> [2,]  0.3838878 1.0000000  0.3612672
#> [3,] -0.2438865 0.3612672  1.0000000
#> 
#> , , 20
#> 
#>            [,1]      [,2]       [,3]
#> [1,]  1.0000000 0.4102077 -0.2619835
#> [2,]  0.4102077 1.0000000  0.3513615
#> [3,] -0.2619835 0.3513615  1.0000000
#> 
#> 
#> $DPCCA
#> , , 1
#> 
#>             [,1]       [,2]       [,3]
#> [1,]  1.00000000 0.05306525 -0.0696031
#> [2,]  0.05306525 1.00000000  0.1539116
#> [3,] -0.06960310 0.15391164  1.0000000
#> 
#> , , 2
#> 
#>             [,1]       [,2]        [,3]
#> [1,]  1.00000000 0.07480503 -0.07532891
#> [2,]  0.07480503 1.00000000  0.19538181
#> [3,] -0.07532891 0.19538181  1.00000000
#> 
#> , , 3
#> 
#>             [,1]       [,2]        [,3]
#> [1,]  1.00000000 0.08563525 -0.08130219
#> [2,]  0.08563525 1.00000000  0.21479570
#> [3,] -0.08130219 0.21479570  1.00000000
#> 
#> , , 4
#> 
#>             [,1]       [,2]        [,3]
#> [1,]  1.00000000 0.09730414 -0.08909513
#> [2,]  0.09730414 1.00000000  0.23336331
#> [3,] -0.08909513 0.23336331  1.00000000
#> 
#> , , 5
#> 
#>             [,1]      [,2]        [,3]
#> [1,]  1.00000000 0.1105059 -0.09822934
#> [2,]  0.11050587 1.0000000  0.25123714
#> [3,] -0.09822934 0.2512371  1.00000000
#> 
#> , , 6
#> 
#>            [,1]      [,2]       [,3]
#> [1,]  1.0000000 0.1254696 -0.1090667
#> [2,]  0.1254696 1.0000000  0.2685402
#> [3,] -0.1090667 0.2685402  1.0000000
#> 
#> , , 7
#> 
#>            [,1]      [,2]       [,3]
#> [1,]  1.0000000 0.1419750 -0.1212738
#> [2,]  0.1419750 1.0000000  0.2853142
#> [3,] -0.1212738 0.2853142  1.0000000
#> 
#> , , 8
#> 
#>            [,1]      [,2]       [,3]
#> [1,]  1.0000000 0.1597116 -0.1346641
#> [2,]  0.1597116 1.0000000  0.3016497
#> [3,] -0.1346641 0.3016497  1.0000000
#> 
#> , , 9
#> 
#>            [,1]      [,2]       [,3]
#> [1,]  1.0000000 0.1783637 -0.1490503
#> [2,]  0.1783637 1.0000000  0.3175485
#> [3,] -0.1490503 0.3175485  1.0000000
#> 
#> , , 10
#> 
#>            [,1]      [,2]       [,3]
#> [1,]  1.0000000 0.1976070 -0.1640976
#> [2,]  0.1976070 1.0000000  0.3327636
#> [3,] -0.1640976 0.3327636  1.0000000
#> 
#> , , 11
#> 
#>            [,1]      [,2]       [,3]
#> [1,]  1.0000000 0.2171873 -0.1796030
#> [2,]  0.2171873 1.0000000  0.3470616
#> [3,] -0.1796030 0.3470616  1.0000000
#> 
#> , , 12
#> 
#>            [,1]      [,2]       [,3]
#> [1,]  1.0000000 0.2562416 -0.2109058
#> [2,]  0.2562416 1.0000000  0.3729279
#> [3,] -0.2109058 0.3729279  1.0000000
#> 
#> , , 13
#> 
#>            [,1]      [,2]       [,3]
#> [1,]  1.0000000 0.2938113 -0.2417564
#> [2,]  0.2938113 1.0000000  0.3964199
#> [3,] -0.2417564 0.3964199  1.0000000
#> 
#> , , 14
#> 
#>            [,1]      [,2]       [,3]
#> [1,]  1.0000000 0.3300192 -0.2723654
#> [2,]  0.3300192 1.0000000  0.4178697
#> [3,] -0.2723654 0.4178697  1.0000000
#> 
#> , , 15
#> 
#>            [,1]      [,2]       [,3]
#> [1,]  1.0000000 0.3646455 -0.3021119
#> [2,]  0.3646455 1.0000000  0.4369441
#> [3,] -0.3021119 0.4369441  1.0000000
#> 
#> , , 16
#> 
#>            [,1]      [,2]       [,3]
#> [1,]  1.0000000 0.3975408 -0.3309304
#> [2,]  0.3975408 1.0000000  0.4537204
#> [3,] -0.3309304 0.4537204  1.0000000
#> 
#> , , 17
#> 
#>            [,1]      [,2]       [,3]
#> [1,]  1.0000000 0.4431623 -0.3721049
#> [2,]  0.4431623 1.0000000  0.4751428
#> [3,] -0.3721049 0.4751428  1.0000000
#> 
#> , , 18
#> 
#>            [,1]      [,2]       [,3]
#> [1,]  1.0000000 0.4844143 -0.4099933
#> [2,]  0.4844143 1.0000000  0.4928105
#> [3,] -0.4099933 0.4928105  1.0000000
#> 
#> , , 19
#> 
#>            [,1]      [,2]       [,3]
#> [1,]  1.0000000 0.5219430 -0.4443264
#> [2,]  0.5219430 1.0000000  0.5079773
#> [3,] -0.4443264 0.5079773  1.0000000
#> 
#> , , 20
#> 
#>            [,1]      [,2]       [,3]
#> [1,]  1.0000000 0.5558792 -0.4756313
#> [2,]  0.5558792 1.0000000  0.5213147
#> [3,] -0.4756313 0.5213147  1.0000000
#> 
#> 
#> $BDFA
#> , , 1
#> 
#>             [,1]
#> [1,]  0.04788416
#> [2,] -0.06903587
#> 
#> , , 2
#> 
#>             [,1]
#> [1,]  0.06873391
#> [2,] -0.07742917
#> 
#> , , 3
#> 
#>             [,1]
#> [1,]  0.07952671
#> [2,] -0.08487699
#> 
#> , , 4
#> 
#>             [,1]
#> [1,]  0.09143237
#> [2,] -0.09425809
#> 
#> , , 5
#> 
#>            [,1]
#> [1,]  0.1051755
#> [2,] -0.1050664
#> 
#> , , 6
#> 
#>            [,1]
#> [1,]  0.1209983
#> [2,] -0.1176421
#> 
#> , , 7
#> 
#>            [,1]
#> [1,]  0.1387593
#> [2,] -0.1316108
#> 
#> , , 8
#> 
#>            [,1]
#> [1,]  0.1582223
#> [2,] -0.1467798
#> 
#> , , 9
#> 
#>            [,1]
#> [1,]  0.1791166
#> [2,] -0.1629505
#> 
#> , , 10
#> 
#>            [,1]
#> [1,]  0.2011081
#> [2,] -0.1797493
#> 
#> , , 11
#> 
#>            [,1]
#> [1,]  0.2238961
#> [2,] -0.1969217
#> 
#> , , 12
#> 
#>            [,1]
#> [1,]  0.2705259
#> [2,] -0.2311165
#> 
#> , , 13
#> 
#>            [,1]
#> [1,]  0.3167721
#> [2,] -0.2639978
#> 
#> , , 14
#> 
#>            [,1]
#> [1,]  0.3623191
#> [2,] -0.2955880
#> 
#> , , 15
#> 
#>            [,1]
#> [1,]  0.4067150
#> [2,] -0.3251089
#> 
#> , , 16
#> 
#>            [,1]
#> [1,]  0.4496748
#> [2,] -0.3524598
#> 
#> , , 17
#> 
#>            [,1]
#> [1,]  0.5104101
#> [2,] -0.3889982
#> 
#> , , 18
#> 
#>            [,1]
#> [1,]  0.5660270
#> [2,] -0.4193775
#> 
#> , , 19
#> 
#>            [,1]
#> [1,]  0.6169858
#> [2,] -0.4436810
#> 
#> , , 20
#> 
#>            [,1]
#> [1,]  0.6629933
#> [2,] -0.4629873
#> 
#> 
#> $BSDFA
#> , , 1
#> 
#>            [,1]       [,2]
#> [1,] 0.05357492 -0.0703432
#> 
#> , , 2
#> 
#>            [,1]        [,2]
#> [1,] 0.07605835 -0.07659403
#> 
#> , , 3
#> 
#>            [,1]        [,2]
#> [1,] 0.08739156 -0.08293942
#> 
#> , , 4
#> 
#>            [,1]        [,2]
#> [1,] 0.09966907 -0.09119014
#> 
#> , , 5
#> 
#>           [,1]       [,2]
#> [1,] 0.1136156 -0.1008628
#> 
#> , , 6
#> 
#>          [,1]       [,2]
#> [1,] 0.129477 -0.1123309
#> 
#> , , 7
#> 
#>           [,1]       [,2]
#> [1,] 0.1470389 -0.1252515
#> 
#> , , 8
#> 
#>           [,1]       [,2]
#> [1,] 0.1659888 -0.1394304
#> 
#> , , 9
#> 
#>           [,1]       [,2]
#> [1,] 0.1859982 -0.1546654
#> 
#> , , 10
#> 
#>           [,1]       [,2]
#> [1,] 0.2067085 -0.1705833
#> 
#> , , 11
#> 
#>           [,1]       [,2]
#> [1,] 0.2278162 -0.1869353
#> 
#> , , 12
#> 
#>          [,1]       [,2]
#> [1,] 0.269952 -0.2197144
#> 
#> , , 13
#> 
#>           [,1]       [,2]
#> [1,] 0.3105386 -0.2517087
#> 
#> , , 14
#> 
#>           [,1]       [,2]
#> [1,] 0.3495212 -0.2829983
#> 
#> , , 15
#> 
#>          [,1]       [,2]
#> [1,] 0.386449 -0.3127448
#> 
#> , , 16
#> 
#>           [,1]       [,2]
#> [1,] 0.4209661 -0.3407491
#> 
#> , , 17
#> 
#>           [,1]       [,2]
#> [1,] 0.4674793 -0.3790964
#> 
#> , , 18
#> 
#>           [,1]       [,2]
#> [1,] 0.5077693 -0.4122088
#> 
#> , , 19
#> 
#>           [,1]       [,2]
#> [1,] 0.5428448 -0.4399985
#> 
#> , , 20
#> 
#>          [,1]       [,2]
#> [1,] 0.572998 -0.4633129
#> 
#> 
#> $UDFA
#> , , 1
#> 
#>          [,1]
#> [1,] 1.923722
#> 
#> , , 2
#> 
#>          [,1]
#> [1,] 3.262179
#> 
#> , , 3
#> 
#>          [,1]
#> [1,] 4.117736
#> 
#> , , 4
#> 
#>          [,1]
#> [1,] 5.113373
#> 
#> , , 5
#> 
#>          [,1]
#> [1,] 6.262682
#> 
#> , , 6
#> 
#>          [,1]
#> [1,] 7.571046
#> 
#> , , 7
#> 
#>          [,1]
#> [1,] 9.045773
#> 
#> , , 8
#> 
#>          [,1]
#> [1,] 10.69611
#> 
#> , , 9
#> 
#>          [,1]
#> [1,] 12.53049
#> 
#> , , 10
#> 
#>          [,1]
#> [1,] 14.55629
#> 
#> , , 11
#> 
#>          [,1]
#> [1,] 16.77631
#> 
#> , , 12
#> 
#>         [,1]
#> [1,] 21.8015
#> 
#> , , 13
#> 
#>          [,1]
#> [1,] 27.59792
#> 
#> , , 14
#> 
#>          [,1]
#> [1,] 34.15056
#> 
#> , , 15
#> 
#>          [,1]
#> [1,] 41.44661
#> 
#> , , 16
#> 
#>          [,1]
#> [1,] 49.47074
#> 
#> , , 17
#> 
#>          [,1]
#> [1,] 62.76324
#> 
#> , , 18
#> 
#>          [,1]
#> [1,] 77.37515
#> 
#> , , 19
#> 
#>          [,1]
#> [1,] 93.09807
#> 
#> , , 20
#> 
#>          [,1]
#> [1,] 109.6879
#> 
#> 
#> $VDFA
#> , , 1
#> 
#>           [,1]      [,2]
#> [1,] 0.1623935 0.1958003
#> 
#> , , 2
#> 
#>           [,1]      [,2]
#> [1,] 0.1199348 0.1500778
#> 
#> , , 3
#> 
#>           [,1]      [,2]
#> [1,] 0.1070124 0.1353335
#> 
#> , , 4
#> 
#>            [,1]      [,2]
#> [1,] 0.09717697 0.1233746
#> 
#> , , 5
#> 
#>            [,1]      [,2]
#> [1,] 0.08947933 0.1133012
#> 
#> , , 6
#> 
#>            [,1]      [,2]
#> [1,] 0.08321413 0.1045084
#> 
#> , , 7
#> 
#>            [,1]       [,2]
#> [1,] 0.07799664 0.09670138
#> 
#> , , 8
#> 
#>            [,1]       [,2]
#> [1,] 0.07356951 0.08972992
#> 
#> , , 9
#> 
#>            [,1]       [,2]
#> [1,] 0.06974123 0.08347583
#> 
#> , , 10
#> 
#>            [,1]       [,2]
#> [1,] 0.06635362 0.07783653
#> 
#> , , 11
#> 
#>            [,1]       [,2]
#> [1,] 0.06328772 0.07271095
#> 
#> , , 12
#> 
#>            [,1]       [,2]
#> [1,] 0.05785637 0.06374575
#> 
#> , , 13
#> 
#>            [,1]       [,2]
#> [1,] 0.05310292 0.05613838
#> 
#> , , 14
#> 
#>            [,1]       [,2]
#> [1,] 0.04882043 0.04956468
#> 
#> , , 15
#> 
#>            [,1]       [,2]
#> [1,] 0.04494315 0.04384751
#> 
#> , , 16
#> 
#>            [,1]       [,2]
#> [1,] 0.04143357 0.03885073
#> 
#> , , 17
#> 
#>            [,1]       [,2]
#> [1,] 0.03675859 0.03246691
#> 
#> , , 18
#> 
#>            [,1]       [,2]
#> [1,] 0.03265476 0.02720074
#> 
#> , , 19
#> 
#>            [,1]       [,2]
#> [1,] 0.02904787 0.02286411
#> 
#> , , 20
#> 
#>            [,1]       [,2]
#> [1,] 0.02586729 0.01929426
#> 
#> 
#> $VDFA2
#> , , 1
#> 
#>          [,1]      [,2]
#> [1,] 0.793504 0.9567396
#> 
#> , , 2
#> 
#>           [,1]     [,2]
#> [1,] 0.8089735 1.012291
#> 
#> , , 3
#> 
#>           [,1]     [,2]
#> [1,] 0.8185995 1.035244
#> 
#> , , 4
#> 
#>          [,1]     [,2]
#> [1,] 0.829658 1.053323
#> 
#> , , 5
#> 
#>           [,1]     [,2]
#> [1,] 0.8419397 1.066088
#> 
#> , , 6
#> 
#>           [,1]     [,2]
#> [1,] 0.8542237 1.072817
#> 
#> , , 7
#> 
#>           [,1]     [,2]
#> [1,] 0.8662782 1.074025
#> 
#> , , 8
#> 
#>           [,1]     [,2]
#> [1,] 0.8779581 1.070812
#> 
#> , , 9
#> 
#>           [,1]    [,2]
#> [1,] 0.8890615 1.06415
#> 
#> , , 10
#> 
#>           [,1]    [,2]
#> [1,] 0.8992921 1.05492
#> 
#> , , 11
#> 
#>           [,1]     [,2]
#> [1,] 0.9083914 1.043646
#> 
#> , , 12
#> 
#>           [,1]     [,2]
#> [1,] 0.9227975 1.016732
#> 
#> , , 13
#> 
#>           [,1]      [,2]
#> [1,] 0.9313457 0.9845832
#> 
#> , , 14
#> 
#>           [,1]      [,2]
#> [1,] 0.9339993 0.9482378
#> 
#> , , 15
#> 
#>           [,1]      [,2]
#> [1,] 0.9324459 0.9097145
#> 
#> , , 16
#> 
#>           [,1]      [,2]
#> [1,] 0.9281461 0.8702883
#> 
#> , , 17
#> 
#>          [,1]      [,2]
#> [1,] 0.917804 0.8106474
#> 
#> , , 18
#> 
#>          [,1]      [,2]
#> [1,] 0.902895 0.7520929
#> 
#> , , 19
#> 
#>          [,1]     [,2]
#> [1,] 0.883985 0.695801
#> 
#> , , 20
#> 
#>           [,1]      [,2]
#> [1,] 0.8616062 0.6426671
#> 
#> 
#> $DMC2
#>              [,1]
#>  [1,] 0.006681850
#>  [2,] 0.009428205
#>  [3,] 0.011482242
#>  [4,] 0.014129292
#>  [5,] 0.017511565
#>  [6,] 0.021865260
#>  [7,] 0.027258163
#>  [8,] 0.033736574
#>  [9,] 0.041311128
#> [10,] 0.049923758
#> [11,] 0.059520973
#> [12,] 0.081113764
#> [13,] 0.104947612
#> [14,] 0.130817102
#> [15,] 0.158163779
#> [16,] 0.186582510
#> [17,] 0.230096876
#> [18,] 0.273399204
#> [19,] 0.315701142
#> [20,] 0.356428536
#> 
#> $R2DFA
#> , , 1
#> 
#>            [,1]
#> [1,] 0.00668185
#> 
#> , , 2
#> 
#>             [,1]
#> [1,] 0.009428205
#> 
#> , , 3
#> 
#>            [,1]
#> [1,] 0.01148224
#> 
#> , , 4
#> 
#>            [,1]
#> [1,] 0.01412929
#> 
#> , , 5
#> 
#>            [,1]
#> [1,] 0.01751156
#> 
#> , , 6
#> 
#>            [,1]
#> [1,] 0.02186526
#> 
#> , , 7
#> 
#>            [,1]
#> [1,] 0.02725816
#> 
#> , , 8
#> 
#>            [,1]
#> [1,] 0.03373657
#> 
#> , , 9
#> 
#>            [,1]
#> [1,] 0.04131113
#> 
#> , , 10
#> 
#>            [,1]
#> [1,] 0.04992376
#> 
#> , , 11
#> 
#>            [,1]
#> [1,] 0.05952097
#> 
#> , , 12
#> 
#>            [,1]
#> [1,] 0.08111376
#> 
#> , , 13
#> 
#>           [,1]
#> [1,] 0.1049476
#> 
#> , , 14
#> 
#>           [,1]
#> [1,] 0.1308171
#> 
#> , , 15
#> 
#>           [,1]
#> [1,] 0.1581638
#> 
#> , , 16
#> 
#>           [,1]
#> [1,] 0.1865825
#> 
#> , , 17
#> 
#>           [,1]
#> [1,] 0.2300969
#> 
#> , , 18
#> 
#>           [,1]
#> [1,] 0.2733992
#> 
#> , , 19
#> 
#>           [,1]
#> [1,] 0.3157011
#> 
#> , , 20
#> 
#>           [,1]
#> [1,] 0.3564285
#> 
#> 
#> $UCIB
#> , , 1
#> 
#>           [,1]
#> [1,] 0.8747325
#> [2,] 0.8388853
#> 
#> , , 2
#> 
#>           [,1]
#> [1,] 0.7869496
#> [2,] 0.7259866
#> 
#> , , 3
#> 
#>           [,1]
#> [1,] 0.7617353
#> [2,] 0.6823126
#> 
#> , , 4
#> 
#>           [,1]
#> [1,] 0.7452674
#> [2,] 0.6424566
#> 
#> , , 5
#> 
#>           [,1]
#> [1,] 0.7362867
#> [2,] 0.6051027
#> 
#> , , 6
#> 
#>           [,1]
#> [1,] 0.7333142
#> [2,] 0.5685608
#> 
#> , , 7
#> 
#>           [,1]
#> [1,] 0.7352797
#> [2,] 0.5325966
#> 
#> , , 8
#> 
#>           [,1]
#> [1,] 0.7413018
#> [2,] 0.4971633
#> 
#> , , 9
#> 
#>           [,1]
#> [1,] 0.7505951
#> [2,] 0.4622732
#> 
#> , , 10
#> 
#>           [,1]
#> [1,] 0.7623528
#> [2,] 0.4281220
#> 
#> , , 11
#> 
#>           [,1]
#> [1,] 0.7758936
#> [2,] 0.3947452
#> 
#> , , 12
#> 
#>           [,1]
#> [1,] 0.8061537
#> [2,] 0.3311123
#> 
#> , , 13
#> 
#>           [,1]
#> [1,] 0.8380652
#> [2,] 0.2719873
#> 
#> , , 14
#> 
#>           [,1]
#> [1,] 0.8706263
#> [2,] 0.2165790
#> 
#> , , 15
#> 
#>           [,1]
#> [1,] 0.9032789
#> [2,] 0.1653649
#> 
#> , , 16
#> 
#>           [,1]
#> [1,] 0.9357537
#> [2,] 0.1182251
#> 
#> , , 17
#> 
#>            [,1]
#> [1,] 0.98301275
#> [2,] 0.05515941
#> 
#> , , 18
#> 
#>             [,1]
#> [1,] 1.027605810
#> [2,] 0.001894893
#> 
#> , , 19
#> 
#>            [,1]
#> [1,]  1.0701558
#> [2,] -0.0416299
#> 
#> , , 20
#> 
#>             [,1]
#> [1,]  1.11056940
#> [2,] -0.07643742
#> 
#> 
#> $LCIB
#> , , 1
#> 
#>            [,1]
#> [1,] -0.7789641
#> [2,] -0.9769571
#> 
#> , , 2
#> 
#>            [,1]
#> [1,] -0.6494817
#> [2,] -0.8808449
#> 
#> , , 3
#> 
#>            [,1]
#> [1,] -0.6026819
#> [2,] -0.8520666
#> 
#> , , 4
#> 
#>            [,1]
#> [1,] -0.5624026
#> [2,] -0.8309728
#> 
#> , , 5
#> 
#>            [,1]
#> [1,] -0.5259356
#> [2,] -0.8152356
#> 
#> , , 6
#> 
#>            [,1]
#> [1,] -0.4913177
#> [2,] -0.8038450
#> 
#> , , 7
#> 
#>            [,1]
#> [1,] -0.4577611
#> [2,] -0.7958181
#> 
#> , , 8
#> 
#>            [,1]
#> [1,] -0.4248571
#> [2,] -0.7907228
#> 
#> , , 9
#> 
#>            [,1]
#> [1,] -0.3923619
#> [2,] -0.7881743
#> 
#> , , 10
#> 
#>            [,1]
#> [1,] -0.3601366
#> [2,] -0.7876206
#> 
#> , , 11
#> 
#>            [,1]
#> [1,] -0.3281014
#> [2,] -0.7885886
#> 
#> , , 12
#> 
#>            [,1]
#> [1,] -0.2651018
#> [2,] -0.7933453
#> 
#> , , 13
#> 
#>            [,1]
#> [1,] -0.2045210
#> [2,] -0.7999829
#> 
#> , , 14
#> 
#>            [,1]
#> [1,] -0.1459882
#> [2,] -0.8077551
#> 
#> , , 15
#> 
#>             [,1]
#> [1,] -0.08984882
#> [2,] -0.81558274
#> 
#> , , 16
#> 
#>             [,1]
#> [1,] -0.03640414
#> [2,] -0.82314462
#> 
#> , , 17
#> 
#>             [,1]
#> [1,]  0.03780755
#> [2,] -0.83315588
#> 
#> , , 18
#> 
#>            [,1]
#> [1,]  0.1044482
#> [2,] -0.8406498
#> 
#> , , 19
#> 
#>            [,1]
#> [1,]  0.1638158
#> [2,] -0.8457320
#> 
#> , , 20
#> 
#>            [,1]
#> [1,]  0.2154172
#> [2,] -0.8495371
#> 
#> 
#> $p.value
#> , , 1
#> 
#>           [,1]
#> [1,] 0.4531469
#> [2,] 0.4385907
#> 
#> , , 2
#> 
#>           [,1]
#> [1,] 0.4222494
#> [2,] 0.4217093
#> 
#> , , 3
#> 
#>           [,1]
#> [1,] 0.4051949
#> [2,] 0.4099331
#> 
#> , , 4
#> 
#>           [,1]
#> [1,] 0.3862835
#> [2,] 0.3957044
#> 
#> , , 5
#> 
#>           [,1]
#> [1,] 0.3647279
#> [2,] 0.3793658
#> 
#> , , 6
#> 
#>           [,1]
#> [1,] 0.3402761
#> [2,] 0.3603842
#> 
#> , , 7
#> 
#>           [,1]
#> [1,] 0.3133350
#> [2,] 0.3391385
#> 
#> , , 8
#> 
#>           [,1]
#> [1,] 0.2845845
#> [2,] 0.3159525
#> 
#> , , 9
#> 
#>           [,1]
#> [1,] 0.2548491
#> [2,] 0.2912564
#> 
#> , , 10
#> 
#>           [,1]
#> [1,] 0.2250462
#> [2,] 0.2657578
#> 
#> , , 11
#> 
#>           [,1]
#> [1,] 0.1960232
#> [2,] 0.2400501
#> 
#> , , 12
#> 
#>           [,1]
#> [1,] 0.1434402
#> [2,] 0.1907203
#> 
#> , , 13
#> 
#>           [,1]
#> [1,] 0.1012492
#> [2,] 0.1470275
#> 
#> , , 14
#> 
#>            [,1]
#> [1,] 0.06958009
#> [2,] 0.11021258
#> 
#> , , 15
#> 
#>            [,1]
#> [1,] 0.04729327
#> [2,] 0.08123192
#> 
#> , , 16
#> 
#>            [,1]
#> [1,] 0.03235573
#> [2,] 0.05947289
#> 
#> , , 17
#> 
#>            [,1]
#> [1,] 0.01925426
#> [2,] 0.03778489
#> 
#> , , 18
#> 
#>            [,1]
#> [1,] 0.01257517
#> [2,] 0.02535634
#> 
#> , , 19
#> 
#>             [,1]
#> [1,] 0.009129903
#> [2,] 0.018450974
#> 
#> , , 20
#> 
#>             [,1]
#> [1,] 0.007380605
#> [2,] 0.014638472
#> 
#> 
#> $TC
#> , , 1
#> 
#>           [,1]
#> [1,] 0.8268483
#> [2,] 0.9079212
#> 
#> , , 2
#> 
#>           [,1]
#> [1,] 0.7182156
#> [2,] 0.8034158
#> 
#> , , 3
#> 
#>           [,1]
#> [1,] 0.6822086
#> [2,] 0.7671896
#> 
#> , , 4
#> 
#>           [,1]
#> [1,] 0.6538350
#> [2,] 0.7367147
#> 
#> , , 5
#> 
#>           [,1]
#> [1,] 0.6311111
#> [2,] 0.7101691
#> 
#> , , 6
#> 
#>           [,1]
#> [1,] 0.6123159
#> [2,] 0.6862029
#> 
#> , , 7
#> 
#>           [,1]
#> [1,] 0.5965204
#> [2,] 0.6642074
#> 
#> , , 8
#> 
#>           [,1]
#> [1,] 0.5830795
#> [2,] 0.6439430
#> 
#> , , 9
#> 
#>           [,1]
#> [1,] 0.5714785
#> [2,] 0.6252237
#> 
#> , , 10
#> 
#>           [,1]
#> [1,] 0.5612447
#> [2,] 0.6078713
#> 
#> , , 11
#> 
#>           [,1]
#> [1,] 0.5519975
#> [2,] 0.5916669
#> 
#> , , 12
#> 
#>           [,1]
#> [1,] 0.5356277
#> [2,] 0.5622288
#> 
#> , , 13
#> 
#>           [,1]
#> [1,] 0.5212931
#> [2,] 0.5359851
#> 
#> , , 14
#> 
#>           [,1]
#> [1,] 0.5083072
#> [2,] 0.5121671
#> 
#> , , 15
#> 
#>           [,1]
#> [1,] 0.4965638
#> [2,] 0.4904738
#> 
#> , , 16
#> 
#>           [,1]
#> [1,] 0.4860789
#> [2,] 0.4706849
#> 
#> , , 17
#> 
#>           [,1]
#> [1,] 0.4726026
#> [2,] 0.4441576
#> 
#> , , 18
#> 
#>           [,1]
#> [1,] 0.4615788
#> [2,] 0.4212724
#> 
#> , , 19
#> 
#>           [,1]
#> [1,] 0.4531700
#> [2,] 0.4020511
#> 
#> , , 20
#> 
#>           [,1]
#> [1,] 0.4475761
#> [2,] 0.3865498
#> 
#> 
#> $VIF
#> , , 1
#> 
#>          [,1]     [,2]
#> [1,] 1.023268 1.023268
#> 
#> , , 2
#> 
#>          [,1]     [,2]
#> [1,] 1.037789 1.037789
#> 
#> , , 3
#> 
#>         [,1]    [,2]
#> [1,] 1.04581 1.04581
#> 
#> , , 4
#> 
#>         [,1]    [,2]
#> [1,] 1.05416 1.05416
#> 
#> , , 5
#> 
#>          [,1]     [,2]
#> [1,] 1.062776 1.062776
#> 
#> , , 6
#> 
#>          [,1]     [,2]
#> [1,] 1.071564 1.071564
#> 
#> , , 7
#> 
#>          [,1]     [,2]
#> [1,] 1.080438 1.080438
#> 
#> , , 8
#> 
#>         [,1]    [,2]
#> [1,] 1.08935 1.08935
#> 
#> , , 9
#> 
#>          [,1]     [,2]
#> [1,] 1.098211 1.098211
#> 
#> , , 10
#> 
#>          [,1]     [,2]
#> [1,] 1.106764 1.106764
#> 
#> , , 11
#> 
#>          [,1]     [,2]
#> [1,] 1.114722 1.114722
#> 
#> , , 12
#> 
#>          [,1]     [,2]
#> [1,] 1.128541 1.128541
#> 
#> , , 13
#> 
#>          [,1]     [,2]
#> [1,] 1.140348 1.140348
#> 
#> , , 14
#> 
#>          [,1]     [,2]
#> [1,] 1.149947 1.149947
#> 
#> , , 15
#> 
#>          [,1]     [,2]
#> [1,] 1.156781 1.156781
#> 
#> , , 16
#> 
#>          [,1]     [,2]
#> [1,] 1.160672 1.160672
#> 
#> , , 17
#> 
#>          [,1]     [,2]
#> [1,] 1.161467 1.161467
#> 
#> , , 18
#> 
#>          [,1]     [,2]
#> [1,] 1.157335 1.157335
#> 
#> , , 19
#> 
#>          [,1]     [,2]
#> [1,] 1.150105 1.150105
#> 
#> , , 20
#> 
#>          [,1]     [,2]
#> [1,] 1.140843 1.140843
#> 
#> 
#> $kappa
#> , , 1
#> 
#>          [,1]
#> [1,] 1.355146
#> 
#> , , 2
#> 
#>          [,1]
#> [1,] 1.471644
#> 
#> , , 3
#> 
#>         [,1]
#> [1,] 1.52938
#> 
#> , , 4
#> 
#>          [,1]
#> [1,] 1.586207
#> 
#> , , 5
#> 
#>          [,1]
#> [1,] 1.642144
#> 
#> , , 6
#> 
#>          [,1]
#> [1,] 1.696971
#> 
#> , , 7
#> 
#>          [,1]
#> [1,] 1.750478
#> 
#> , , 8
#> 
#>          [,1]
#> [1,] 1.802666
#> 
#> , , 9
#> 
#>          [,1]
#> [1,] 1.853253
#> 
#> , , 10
#> 
#>          [,1]
#> [1,] 1.901025
#> 
#> , , 11
#> 
#>          [,1]
#> [1,] 1.944657
#> 
#> , , 12
#> 
#>          [,1]
#> [1,] 2.018826
#> 
#> , , 13
#> 
#>          [,1]
#> [1,] 2.080811
#> 
#> , , 14
#> 
#>         [,1]
#> [1,] 2.13039
#> 
#> , , 15
#> 
#>          [,1]
#> [1,] 2.165292
#> 
#> , , 16
#> 
#>          [,1]
#> [1,] 2.185026
#> 
#> , , 17
#> 
#>          [,1]
#> [1,] 2.189049
#> 
#> , , 18
#> 
#>          [,1]
#> [1,] 2.168109
#> 
#> , , 19
#> 
#>        [,1]
#> [1,] 2.1312
#> 
#> , , 20
#> 
#>          [,1]
#> [1,] 2.083381
#> 
#> 
#> $R2adj
#> , , 1
#> 
#>               [,1]
#> [1,] -7.160626e-06
#> 
#> , , 2
#> 
#>             [,1]
#> [1,] 0.002757688
#> 
#> , , 3
#> 
#>             [,1]
#> [1,] 0.004825557
#> 
#> , , 4
#> 
#>             [,1]
#> [1,] 0.007490432
#> 
#> , , 5
#> 
#>            [,1]
#> [1,] 0.01089548
#> 
#> , , 6
#> 
#>            [,1]
#> [1,] 0.01527849
#> 
#> , , 7
#> 
#>            [,1]
#> [1,] 0.02070771
#> 
#> , , 8
#> 
#>            [,1]
#> [1,] 0.02722975
#> 
#> , , 9
#> 
#>            [,1]
#> [1,] 0.03485531
#> 
#> , , 10
#> 
#>            [,1]
#> [1,] 0.04352594
#> 
#> , , 11
#> 
#>            [,1]
#> [1,] 0.05318778
#> 
#> , , 12
#> 
#>            [,1]
#> [1,] 0.07492598
#> 
#> , , 13
#> 
#>            [,1]
#> [1,] 0.09892032
#> 
#> , , 14
#> 
#>          [,1]
#> [1,] 0.124964
#> 
#> , , 15
#> 
#>           [,1]
#> [1,] 0.1524948
#> 
#> , , 16
#> 
#>          [,1]
#> [1,] 0.181105
#> 
#> , , 17
#> 
#>           [,1]
#> [1,] 0.2249123
#> 
#> , , 18
#> 
#>           [,1]
#> [1,] 0.2685063
#> 
#> , , 19
#> 
#>           [,1]
#> [1,] 0.3110931
#> 
#> , , 20
#> 
#>           [,1]
#> [1,] 0.3520947
#> 
#>